Split envoy package into API and Server
